requirements.txt
aniso8601==9.0.1
boto3==1.20.37
botocore==1.23.37
click==8.0.3
colorama==0.4.4
Flask==1.1.2
Flask-RESTful==0.3.9
flask-restful-swagger-2==0.35
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==2.0.2
